move : admin 엔티티, 레포지토리 위치 변경
