+++ backend/support/domain/src/main/java/kr/co/yigil/admin/domain/Admin.java
@@ -1,4 +1,4 @@
-package kr.co.yigil.admin.domain.admin;
+package kr.co.yigil.admin.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.ElementCollection;
@@ -12,7 +12,6 @@ import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
