+++ backend/support/domain/src/main/java/kr/co/yigil/admin/domain/AdminSignUp.java
@@ -1,4 +1,5 @@
-package kr.co.yigil.admin.domain.adminSignUp;
+package kr.co.yigil.admin.domain;
+
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
