+++ backend/support/domain/src/main/java/kr/co/yigil/admin/infrastructure/AdminSignUpRepository.java
@@ -1,6 +1,6 @@
-package kr.co.yigil.admin.infrastructure.adminSignUp;
+package kr.co.yigil.admin.infrastructure;
 
-import kr.co.yigil.admin.domain.adminSignUp.AdminSignUp;
+import kr.co.yigil.admin.domain.AdminSignUp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdminSignUpRepository extends JpaRepository<AdminSignUp, Long> {
