+++ backend/yigil-admin/src/main/java/kr/co/yigil/admin/application/AdminFacade.java
@@ -1,13 +1,11 @@
 package kr.co.yigil.admin.application;
 
+import kr.co.yigil.admin.domain.AdminSignUp;
 import kr.co.yigil.admin.domain.admin.AdminCommand.AdminPasswordUpdateRequest;
-import kr.co.yigil.admin.domain.admin.AdminCommand.AdminUpdateRequest;
 import kr.co.yigil.admin.domain.admin.AdminCommand.LoginRequest;
-import kr.co.yigil.admin.domain.admin.AdminInfo;
 import kr.co.yigil.admin.domain.admin.AdminInfo.AdminDetailInfoResponse;
 import kr.co.yigil.admin.domain.admin.AdminInfo.AdminInfoResponse;
 import kr.co.yigil.admin.domain.admin.AdminService;
-import kr.co.yigil.admin.domain.adminSignUp.AdminSignUp;
 import kr.co.yigil.admin.domain.adminSignUp.AdminSignUpCommand.AdminSignUpRequest;
 import kr.co.yigil.admin.domain.adminSignUp.AdminSignUpService;
 import kr.co.yigil.admin.interfaces.dto.request.AdminSignUpListRequest;
@@ -53,10 +51,6 @@ public class AdminFacade {
     public AdminDetailInfoResponse getAdminDetailInfoByEmail(String email) {
         return adminService.getAdminDetailInfoByEmail(email);
     }
-
-//    public void updateAdminDetailInfo(String email, AdminUpdateRequest command) {
-//        adminService.updateAdminDetailInfo(email, command);
-//    }
 
     public void testSignUp() {
         adminService.testSignUp();
