+++ backend/yigil-admin/src/main/java/kr/co/yigil/admin/domain/admin/AdminStore.java
@@ -1,5 +1,7 @@
 package kr.co.yigil.admin.domain.admin;
 
+import kr.co.yigil.admin.domain.Admin;
+
 public interface AdminStore {
 
     void store(Admin admin);
