+++ backend/yigil-admin/src/main/java/kr/co/yigil/admin/domain/adminSignUp/AdminSignUpServiceImpl.java
@@ -4,7 +4,8 @@ import static kr.co.yigil.global.exception.ExceptionCode.ADMIN_ALREADY_EXISTED;
 
 import java.util.ArrayList;
 import java.util.List;
-import kr.co.yigil.admin.domain.admin.Admin;
+import kr.co.yigil.admin.domain.Admin;
+import kr.co.yigil.admin.domain.AdminSignUp;
 import kr.co.yigil.admin.domain.admin.AdminReader;
 import kr.co.yigil.admin.domain.admin.AdminStore;
 import kr.co.yigil.admin.domain.adminSignUp.AdminSignUpCommand.AdminSignUpRequest;
