+++ backend/yigil-admin/src/main/java/kr/co/yigil/admin/domain/adminSignUp/AdminSignUpStore.java
@@ -1,5 +1,7 @@
 package kr.co.yigil.admin.domain.adminSignUp;
 
+import kr.co.yigil.admin.domain.AdminSignUp;
+
 public interface AdminSignUpStore {
 
     void store(AdminSignUp entity);
