+++ backend/yigil-admin/src/main/java/kr/co/yigil/admin/infrastructure/admin/AdminReaderImpl.java
@@ -2,8 +2,9 @@ package kr.co.yigil.admin.infrastructure.admin;
 
 import static kr.co.yigil.global.exception.ExceptionCode.ADMIN_NOT_FOUND;
 
-import kr.co.yigil.admin.domain.admin.Admin;
+import kr.co.yigil.admin.domain.Admin;
 import kr.co.yigil.admin.domain.admin.AdminReader;
+import kr.co.yigil.admin.infrastructure.AdminRepository;
 import kr.co.yigil.global.exception.BadRequestException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
