+++ backend/yigil-admin/src/main/java/kr/co/yigil/admin/infrastructure/admin/AdminStoreImpl.java
@@ -1,7 +1,8 @@
 package kr.co.yigil.admin.infrastructure.admin;
 
-import kr.co.yigil.admin.domain.admin.Admin;
+import kr.co.yigil.admin.domain.Admin;
 import kr.co.yigil.admin.domain.admin.AdminStore;
+import kr.co.yigil.admin.infrastructure.AdminRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
