+++ backend/yigil-admin/src/main/java/kr/co/yigil/admin/interfaces/dto/request/AdminSignupRequest.java
@@ -1,6 +1,5 @@
 package kr.co.yigil.admin.interfaces.dto.request;
 
-import kr.co.yigil.admin.domain.adminSignUp.AdminSignUp;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
