+++ backend/yigil-admin/src/main/java/kr/co/yigil/admin/interfaces/dto/response/AdminInfoResponse.java
@@ -1,6 +1,5 @@
 package kr.co.yigil.admin.interfaces.dto.response;
 
-import kr.co.yigil.admin.domain.admin.Admin;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
