+++ backend/yigil-admin/src/main/java/kr/co/yigil/auth/application/CustomUserDetailsService.java
@@ -2,8 +2,8 @@ package kr.co.yigil.auth.application;
 
 import static kr.co.yigil.global.exception.ExceptionCode.ADMIN_NOT_FOUND;
 
-import kr.co.yigil.admin.domain.admin.Admin;
-import kr.co.yigil.admin.infrastructure.admin.AdminRepository;
+import kr.co.yigil.admin.domain.Admin;
+import kr.co.yigil.admin.infrastructure.AdminRepository;
 import kr.co.yigil.global.exception.BadRequestException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
