--- backend/yigil-admin/src/main/java/kr/co/yigil/file/domain/FileType.java
@@ -1,6 +1,0 @@
-package kr.co.yigil.file.domain;
-
-public enum FileType {
-    IMAGE,
-    VIDEO
-}
