+++ backend/yigil-admin/src/main/java/kr/co/yigil/file/domain/FileUploader.java
@@ -1,9 +1,10 @@
 package kr.co.yigil.file.domain;
 
+import kr.co.yigil.file.AttachFile;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface FileUploader {
 
-    AdminAttachFile upload(MultipartFile file);
+    AttachFile upload(MultipartFile file);
 
 }
