+++ backend/yigil-admin/src/test/java/kr/co/yigil/admin/infrastructure/adminSignUp/AdminSignUpStoreImplTest.java
@@ -3,7 +3,8 @@ package kr.co.yigil.admin.infrastructure.adminSignUp;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
-import kr.co.yigil.admin.domain.adminSignUp.AdminSignUp;
+import kr.co.yigil.admin.domain.AdminSignUp;
+import kr.co.yigil.admin.infrastructure.AdminSignUpRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
